requirements.txt
streamlit
pandas
openpyxl
google-generativeai
scikit-learn
python-dotenv
